Migrate assistant/builder routes to Hono
