+++ front-api/middleware/utils.ts
@@ -1,5 +1,21 @@
 import type { Context } from "hono";
+import type { ContentfulStatusCode } from "hono/utils/http-status";
 import type { NextApiRequest, NextApiResponse } from "next";
+
+import type { APIErrorWithStatusCode } from "@app/types/error";
+
+/**
+ * Returns a JSON response built from an `APIErrorWithStatusCode`. The
+ * `status_code` field is typed as `number` but Hono's c.json expects the
+ * narrower `ContentfulStatusCode` — the cast is safe because all our status
+ * codes are valid HTTP error codes.
+ */
+export function jsonApiError(c: Context, err: APIErrorWithStatusCode) {
+  return c.json(
+    { error: err.api_error },
+    err.status_code as ContentfulStatusCode
+  );
+}
 
 export function parseCookieHeader(
   header: string | undefined
