+++ front-api/routes/w/assistant/index.ts
@@ -1,0 +1,9 @@
+import { Hono } from "hono";
+
+import { builderApp } from "./builder";
+
+// Mounted at /api/w/:wId/assistant. workspaceAuth is applied by the parent
+// workspace sub-app.
+export const assistantApp = new Hono();
+
+assistantApp.route("/builder", builderApp);
