+++ front/lib/api/assistant/conversation/helper.ts
@@ -1,6 +1,7 @@
 import { apiError } from "@app/logger/withlogging";
 import type { ConversationErrorType } from "@app/types/assistant/conversation";
 import { ConversationError } from "@app/types/assistant/conversation";
+import type { APIErrorWithStatusCode } from "@app/types/error";
 import { isOverflowingDBString } from "@app/types/shared/utils/general";
 import type { NextApiRequest, NextApiResponse } from "next";
 
@@ -15,33 +16,40 @@ const STATUS_FOR_ERROR_TYPE: Record<ConversationErrorType, number> = {
   conversation_context_usage_not_found: 404,
 };
 
-export function apiErrorForConversation(
-  req: NextApiRequest,
-  res: NextApiResponse,
-  error: Error
-) {
+/**
+ * Maps a conversation error to the standard `{ status_code, api_error }`
+ * shape. Use this from any framework (Next or Hono) — only the response
+ * dispatch differs.
+ */
+export function getConversationApiError(error: Error): APIErrorWithStatusCode {
   if (error instanceof ConversationError) {
-    return apiError(req, res, {
+    return {
       status_code: STATUS_FOR_ERROR_TYPE[error.type],
       api_error: {
         type: error.type,
         message: error.message,
       },
-    });
+    };
   }
-
-  return apiError(
-    req,
-    res,
-    {
-      status_code: 500,
-      api_error: {
-        type: "internal_server_error",
-        message: "An internal server error occurred.",
-      },
+  return {
+    status_code: 500,
+    api_error: {
+      type: "internal_server_error",
+      message: "An internal server error occurred.",
     },
-    error
-  );
+  };
+}
+
+export function apiErrorForConversation(
+  req: NextApiRequest,
+  res: NextApiResponse,
+  error: Error
+) {
+  const apiErr = getConversationApiError(error);
+  if (error instanceof ConversationError) {
+    return apiError(req, res, apiErr);
+  }
+  return apiError(req, res, apiErr, error);
 }
 
 export function isUserMessageContextOverflowing(context: {
