+++ front/pages/api/w/[wId]/assistant/builder/sidekick/prompt/existing.ts
@@ -1,202 +1,13 @@
 /** @ignoreswagger */
-import { getAgentConfiguration } from "@app/lib/api/assistant/configuration/agent";
-import type { AgentMessageFeedbackWithMetadataType } from "@app/lib/api/assistant/feedback";
-import { getAgentFeedbacks } from "@app/lib/api/assistant/feedback";
-import { fetchAgentOverview } from "@app/lib/api/assistant/observability/overview";
-import { buildAgentAnalyticsBaseQuery } from "@app/lib/api/assistant/observability/utils";
+// @migration-status: MIGRATED_TO_HONO
+// @migration-target: front-api/routes/w/assistant/builder/sidekick.ts
+import { buildExistingAgentPrompt } from "@app/lib/api/assistant/builder/sidekick_prompts";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
 import type { Authenticator } from "@app/lib/auth";
-import { AgentSuggestionResource } from "@app/lib/resources/agent_suggestion_resource";
-import logger from "@app/logger/logger";
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types/error";
 import { isString } from "@app/types/shared/utils/general";
 import type { NextApiRequest, NextApiResponse } from "next";
-
-const FEEDBACK_LIMIT = 50;
-const OLDER_FEEDBACK_LIMIT = 10;
-const OLDER_FEEDBACK_MAX_AGE_MS = 90 * 24 * 60 * 60 * 1000; // 3 months
-const INSIGHTS_DAYS = 30;
-const MAX_PENDING_SUGGESTIONS_IN_FIRST_MESSAGE = 20;
-
-function buildFirstMessage({
-  feedbackMarkdown,
-  insightsMarkdown,
-  pendingSuggestions,
-}: {
-  feedbackMarkdown: string | null;
-  insightsMarkdown: string | null;
-  pendingSuggestions: Array<{ sId: string; kind: string }>;
-}): string {
-  const dataSections = [feedbackMarkdown, insightsMarkdown]
-    .filter(Boolean)
-    .join("\n\n");
-
-  const suggestionDirectives =
-    pendingSuggestions.length > 0 &&
-    pendingSuggestions
-      .map((s) => `:agent_suggestion[]{sId=${s.sId} kind=${s.kind}}`)
-      .join("\n");
-
-  const pendingSuggestionsSection = suggestionDirectives
-    ? `
-<pending_suggestions>
-Output these directives verbatim so the suggestion cards render:
-${suggestionDirectives}
-</pending_suggestions>`
-    : "";
-
-  return `<dust_system>
-This is an existing agent.
-
-## Opening message
-Do NOT call \`get_agent_config\` in this first message. Based only on the information provided in <existing_agent_data_section>:
-- If pending suggestions exist (see <pending_suggestions> below), output their directives to render them as cards.
-- If negative feedback patterns exist in the current agent version, mention it as the top issue. Feedback from previous versions are provided for reference, but should not be mentioned in the opening message.
-
-In addition, ask the user if you should suggest additional improvements or if there is something specific they'd like to work on.
-Keep the first message to 1–2 sentences (plus any suggestion cards from <pending_suggestions>). Response must fit in the sidekick panel without scrolling.
-Do not make assumptions about the users's intent. Given that this is an existing agent, the user is likely to be asking for specific improvements or to work on a specific issue.
-
-<existing_agent_data_section>
-${dataSections ? `\n${dataSections}\n` : ""}
-</existing_agent_data_section>
-${pendingSuggestionsSection}
-</dust_system>`;
-}
-
-function formatFeedbackItem(f: AgentMessageFeedbackWithMetadataType): string {
-  const direction = f.thumbDirection === "up" ? "POSITIVE" : "NEGATIVE";
-  const content = f.content ? `: ${f.content}` : "";
-  return `- [${direction}]${content}`;
-}
-
-function appendFeedbackSection(
-  lines: string[],
-  title: string,
-  items: AgentMessageFeedbackWithMetadataType[]
-): void {
-  if (items.length === 0) {
-    return;
-  }
-  const positive = items.filter((f) => f.thumbDirection === "up").length;
-  const negative = items.filter((f) => f.thumbDirection === "down").length;
-  lines.push(
-    "",
-    `${title} (${items.length} total, ${positive} positive, ${negative} negative):`
-  );
-  for (const f of items) {
-    lines.push(formatFeedbackItem(f));
-  }
-}
-
-async function fetchFeedbackMarkdown(
-  auth: Authenticator,
-  agentConfigurationId: string
-): Promise<string | null> {
-  const agentConfiguration = await getAgentConfiguration(auth, {
-    agentId: agentConfigurationId,
-    variant: "light",
-  });
-
-  if (!agentConfiguration) {
-    return null;
-  }
-
-  const currentVersion = agentConfiguration.version;
-
-  const feedbacksRes = await getAgentFeedbacks({
-    auth,
-    agentConfigurationId,
-    withMetadata: true,
-    paginationParams: {
-      limit: FEEDBACK_LIMIT,
-      orderColumn: "id",
-      orderDirection: "desc",
-    },
-    filter: "active",
-  });
-
-  if (feedbacksRes.isErr()) {
-    logger.warn(
-      { err: feedbacksRes.error },
-      "Failed to fetch feedback for sidekick first message"
-    );
-    return null;
-  }
-
-  const feedbacks = feedbacksRes.value.filter(
-    (f): f is AgentMessageFeedbackWithMetadataType => true
-  );
-
-  if (feedbacks.length === 0) {
-    return null;
-  }
-
-  const latestVersionFeedback = feedbacks.filter(
-    (f) => f.agentConfigurationVersion === currentVersion
-  );
-
-  const cutoffMs = Date.now() - OLDER_FEEDBACK_MAX_AGE_MS;
-  const olderFeedback = feedbacks
-    .filter(
-      (f) =>
-        f.agentConfigurationVersion !== currentVersion &&
-        new Date(f.createdAt).getTime() >= cutoffMs
-    )
-    .slice(0, OLDER_FEEDBACK_LIMIT);
-
-  if (latestVersionFeedback.length === 0 && olderFeedback.length === 0) {
-    return null;
-  }
-
-  const lines = ["<feedback>"];
-
-  appendFeedbackSection(
-    lines,
-    `Current version (v${currentVersion})`,
-    latestVersionFeedback
-  );
-  appendFeedbackSection(lines, "Previous versions", olderFeedback);
-
-  lines.push("</feedback>");
-  return lines.join("\n");
-}
-
-async function fetchInsightsMarkdown(
-  auth: Authenticator,
-  agentConfigurationId: string
-): Promise<string | null> {
-  const owner = auth.getNonNullableWorkspace();
-  const baseQuery = buildAgentAnalyticsBaseQuery({
-    workspaceId: owner.sId,
-    agentId: agentConfigurationId,
-    days: INSIGHTS_DAYS,
-  });
-
-  const overviewResult = await fetchAgentOverview(baseQuery, INSIGHTS_DAYS);
-
-  if (overviewResult.isErr()) {
-    logger.warn(
-      { err: overviewResult.error },
-      "Failed to fetch insights for sidekick first message"
-    );
-    return null;
-  }
-
-  const o = overviewResult.value;
-  const lines = [
-    "<insights>",
-    `Period: last ${INSIGHTS_DAYS} days`,
-    `Active users: ${o.activeUsers}`,
-    `Conversations: ${o.conversationCount}`,
-    `Messages: ${o.messageCount}`,
-    `Feedback: ${o.positiveFeedbacks} positive, ${o.negativeFeedbacks} negative`,
-    "</insights>",
-  ];
-
-  return lines.join("\n");
-}
 
 async function handler(
   req: NextApiRequest,
@@ -218,30 +29,9 @@ async function handler(
         });
       }
 
-      const [feedbackMarkdown, insightsMarkdown, pendingSuggestions] =
-        await Promise.all([
-          fetchFeedbackMarkdown(auth, agentConfigurationId),
-          fetchInsightsMarkdown(auth, agentConfigurationId),
-          AgentSuggestionResource.listByAgentConfigurationId(
-            auth,
-            agentConfigurationId,
-            {
-              states: ["pending"],
-              limit: MAX_PENDING_SUGGESTIONS_IN_FIRST_MESSAGE,
-            }
-          ),
-        ]);
-
-      return res.status(200).json(
-        buildFirstMessage({
-          feedbackMarkdown,
-          insightsMarkdown,
-          pendingSuggestions: pendingSuggestions.map((s) => ({
-            sId: s.sId,
-            kind: s.kind,
-          })),
-        })
-      );
+      return res
+        .status(200)
+        .json(await buildExistingAgentPrompt(auth, agentConfigurationId));
     }
     default:
       return apiError(req, res, {
