+++ front/pages/api/w/[wId]/assistant/builder/suggestions.ts
@@ -1,4 +1,6 @@
 /** @ignoreswagger */
+// @migration-status: MIGRATED_TO_HONO
+// @migration-target: front-api/routes/w/assistant/builder.ts
 import { getBuilderSuggestions } from "@app/lib/api/assistant/suggestions";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
 import type { Authenticator } from "@app/lib/auth";
